refactor(automation): make agent stage provider selectable
